+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+* Fix deadlock in `crud.schema()` after schema reload error (#479).
+
 ## [1.7.1] - 16-01-26
 
 ### Changed